requirements.txt
python-dotenv
pyodbc
requests
google-generativeai